the last one

+++ BadTripPS/BadTripPS/instruction.cpp
@@ -29,7 +29,7 @@ uint32_t _instruction::ImmValue() {
 
 // like a imm_se
 uint32_t _instruction::SignExt() {
-	int16_t v = (data & 0xffff);
+	int16_t v = (int16_t)(data & 0xffff);
 	return (uint32_t)v;
 };
 
@@ -48,6 +48,7 @@ uint32_t _instruction::ImmJump() {
 	return (data & 0x3ffffff);
 };
 
+//like a s
 pscx_memory::_regIndex _instruction::s() {
 	return pscx_memory::_regIndex((data >> 21) & 0x1f);
 };
